+++ src-web/hooks/usePreferredAppearance.ts
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react';
 import type { Appearance } from '../lib/theme/appearance';
-import { getCSSAppearance } from '../lib/theme/appearance';
-import { subscribeToPreferredAppearance } from '../lib/theme/window';
+import { getCSSAppearance, subscribeToPreferredAppearance } from '../lib/theme/appearance';
 
 export function usePreferredAppearance() {
   const [preferredAppearance, setPreferredAppearance] = useState<Appearance>(getCSSAppearance());
